+++ examples/with_writer.rs
@@ -1,5 +1,0 @@
-fn main() {
-    let file_appender = tracing_appender::rolling::hourly("./", "test.log");
-    traceon::builder().writer(file_appender).on();
-    tracing::info!("wow cool!");
-}